requirements.txt
autopep8==1.6.0
blinker==1.4
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
SQLAlchemy==1.4.31
toml==0.10.2
Werkzeug==2.0.3
